requirements.txt
wikipedia==1.4.0
python-dotenv~=0.19.2
aiogram~=2.23.1